pyproject.toml
[project]
name = "gemini-transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform>=1.85.0",
    "rich>=13.9.4",
]
